ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
python-dotenv
ffmpeg-python
langchain-groq
langchain-core
groq
gTTS
